modify folder and files
