--- lib/github_stalking/issues.ex
@@ -1,7 +1,0 @@
-defmodule GithubStalking.Issues do
-  @moduledoc"""
-  """
-
-  @derive [Poison.Encoder]
-  defstruct [:repo_full_path, :numbers]
-end
