+++ lib/github_stalking/slack/slack.ex
