+++ test/issue_test.exs
@@ -1,4 +1,4 @@
-defmodule GithubStalking.IssueSpecifierTest do
+defmodule GithubStalking.IssueTest do
   use ExUnit.Case
 
   setup_all do
@@ -12,24 +12,24 @@ defmodule GithubStalking.IssueSpecifierTest do
   end
 
   test "issue 1 should be updated after 2016-02-13T01:05:18Z" do
-    pre_issue = %GithubStalking.Issue{number: 22, updated_at: "2016-02-13T01:05:18Z"}
-    pre_issues = %{22 => pre_issue} 
+    pre_issue = %GithubStalking.Github.Issue{number: 1, updated_at: "2016-03-03T01:05:18Z"}
+    pre_issues = %{1=> pre_issue} 
     
-    response = GithubStalking.IssueSpecifier.updated_open_issues("letusfly85", "github_stalking", pre_issues)
+    response = GithubStalking.Github.Issue.updated_open_issues("letusfly85", "github_stalking_test", pre_issues)
     case response do
       {:ok, issues} -> 
-        assert length(Enum.sort(issues)) == 6
-        assert hd(Enum.sort(issues))["title"] == "notify to slack"
+        assert length(Enum.sort(issues)) == 2
+        assert hd(Enum.sort(issues))["title"] == "test issue 1"
       {:error, _}   -> raise("connection to GitHub is refused!!!")
     end
 
   end
 
   test "issue 6 should be updated after 2016-02-13T01:05:18Z" do
-    pre_issue = %GithubStalking.Issue{number: 6}
+    pre_issue = %GithubStalking.Github.Issue{number: 6}
     pre_issues = %{6 => pre_issue} 
 
-    response = GithubStalking.IssueSpecifier.closed_issues("letusfly85", "github_stalking", pre_issues)
+    response = GithubStalking.Github.Issue.closed_issues("letusfly85", "github_stalking_test", pre_issues)
     case response do
       {:ok, issues} -> 
         assert length(issues) >= 1
